Working prototype
It now works with audio volume , but it only goes from no vilume to
full folume. need it to ramp back down when we pass the perimeter for
the sound, ie: going past the lightyear distance where the sound ends
and an new one begins.

Also the click to travel text should only be visible in the middle of
the screen until you have dragged the mouse.

Moved the music inside ./data/audio

Removed the startext inside the rectangle. too much text already!